Create HackerRank_C++_3
Input Format

You are given two strings, a and b, separated by a new line. Each string will consist of lower case Latin characters ('a'-'z').

Output Format

In the first line print two space-separated integers, representing the length of a and b respectively.
In the second line print the string produced by concatenating a and b,  (a+b).
In the third line print two strings separated by a space, a1 and b1 . and are the same, respectively, except that their first characters are swapped.